+++ src/index.js
@@ -5,11 +5,12 @@ import './js/pagination';
 import './getMoviesForKeyword';
 import { fonNightDay } from './fonNightDay';
 import { modalFormLogin } from './modalFormLogin';
-
+import { initialLoader } from './js/loader';
 
 
 fonNightDay();
 modalFormLogin();
+initialLoader();
 
 // import './fetchMain';
